Cleanup core package
* move optimization that uses diagnostics to experimental subpackage
* fix bug in adagrad_optimize
