+++ viabel/experimental/optimization_diagnostics.py
